+++ data/index.js
@@ -14,6 +14,7 @@ export * from './StoreSelectionModel';
 export * from './UrlStore';
 
 export * from './cube/aggregate/Aggregator';
+export * from './cube/aggregate/AverageAggregator';
 export * from './cube/aggregate/MaxAggregator';
 export * from './cube/aggregate/MinAggregator';
 export * from './cube/aggregate/NullAggregator';